Relocate odin command
