+++ cmd/odin/cmd/snapshot_list.go
@@ -6,37 +6,32 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/poka-yoke/spaceflight/mcc/odin/odin"
+	"github.com/poka-yoke/spaceflight/pkg/odin"
 )
 
-// snapshotCreateCmd represents the snapshot create command
-var snapshotCreateCmd = &cobra.Command{
-	Use:   "create instanceID snapshotID",
-	Short: "Creates a new snapshot",
-	Long: `Create a new snapshot of the specified instance,
-using the specified snapshot name.`,
+// snapshotListCmd represents the snapshot list command
+var snapshotListCmd = &cobra.Command{
+	Use:   "list identifier",
+	Short: "Lists all snapshots",
+	Long:  `Lists all snapshots, ordered from newer to older.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		if len(args) != 2 {
-			log.Fatal(CreateParamsReq)
+		if len(args) != 1 {
+			args = append(args, "")
 		}
 		svc := odin.Init()
-		snapshot, err := odin.CreateSnapshot(
+		snapshots, err := odin.ListSnapshots(
 			args[0],
-			args[1],
 			svc,
 		)
 		if err != nil {
 			log.Fatalf("Error: %s", err)
 		}
-		fmt.Printf(
-			"%s snapshot is being created",
-			*snapshot.DBSnapshotIdentifier,
-		)
+		fmt.Println(odin.PrintSnapshots(snapshots))
 	},
 }
 
 func init() {
-	SnapshotCmd.AddCommand(snapshotCreateCmd)
+	SnapshotCmd.AddCommand(snapshotListCmd)
 
 	// Here you will define your flags and configuration settings.
