+++ cmd/odin/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/poka-yoke/spaceflight/cmd/odin/cmd"
+
+func main() {
+	cmd.Execute()
+}
