--- mcc/odin/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/poka-yoke/spaceflight/mcc/odin/cmd"
-
-func main() {
-	cmd.Execute()
-}
