+++ pkg/odin/instance_delete.go
